+++ src/assembles/Util.h
@@ -8,7 +8,7 @@ extern void copy_bit(uint32_t *in, uint32_t change, uint8_t start, uint8_t end);
 // check in 32 bit or 64 bit;
 extern bool check_bit(char *in);
 
-extern char *substring(char *str, uint8_t start, uint8_t end);
+extern char *substring(char *str, uint32_t start, uint32_t end);
 
 extern int to_int(char *in);
